base/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pytz>=2023.3